Remove unused mouse keys
This commit removes the unused mouse key definitions and the associated
functionality tied to the `TG_SCRL` keycode from the Gherkin keyboard
layouts. The changes are reflected in multiple files including
`placek.c`, `placek.h`, and the keymap files for both the `placek` and
`rev_placek` layouts. This cleanup allows for a more streamlined
codebase and ensures that only relevant key functionalities are
retained, improving the overall readability and maintainability of the
code. Users will benefit from a cleaner layout with no extra,
unintuitive keys associated with mouse functionality.